requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
crewai==0.22.3
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
docstring-parser==0.15
exceptiongroup==1.2.2
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.3.0
googleapis-common-protos==1.69.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
instructor==0.5.2
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.0.5
langchain-text-splitters==0.0.2
langsmith==0.1.147
litellm==1.63.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ollama==0.4.7
openai==1.65.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
orjson==3.10.15
packaging==23.2
pluggy==1.5.0
propcache==0.3.0
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.5
python-dotenv==1.0.0
PyYAML==6.0.2
referencing==0.36.2
regex==2023.12.25
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
watchdog==6.0.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
